requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0

# AI/ML Stack (xAI via OpenAI SDK)
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.19
sentence-transformers>=2.2.0

# Vector Database
chromadb>=0.4.0

# Hybrid Search
rank_bm25>=0.2.2

# UI
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0
tenacity>=8.2.0
httpx>=0.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
